Simplify DPC integration tests
