+++ .integration/src/lib.rs
@@ -13,9 +13,3 @@
 
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
-
-pub mod testnet1;
-pub use testnet1::*;
-
-pub mod testnet2;
-pub use testnet2::*;
