--- .integration/src/testnet1.rs
@@ -1,29 +1,0 @@
-// Copyright (C) 2019-2021 Aleo Systems Inc.
-// This file is part of the snarkVM library.
-
-// The snarkVM library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The snarkVM library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
-
-use snarkvm_dpc::{testnet1::*, DPCScheme};
-
-use rand::{CryptoRng, Rng};
-
-pub fn setup_or_load_dpc<R: Rng + CryptoRng>(verify_only: bool, rng: &mut R) -> Testnet1DPC {
-    match Testnet1DPC::load(verify_only) {
-        Ok(dpc) => dpc,
-        Err(err) => {
-            println!("error - {}, re-running parameter Setup", err);
-            Testnet1DPC::setup(rng).expect("DPC setup failed")
-        }
-    }
-}
